cleanup electron
