+++ electron/src/helpers/controllers/b-device-controller.ts
