+++ electron/src/helpers/kk-state-controller/firmwareUtils.ts
