+++ electron/src/helpers/kk-state-controller/index.ts
