+++ electron/src/helpers/kk-state-controller/types.ts
