+++ electron/src/helpers/kk-state-controller/walletUtils.ts
